add map service tests
The service will keep a cache of map elements based on location. If the
map element exists in the cache, the service will return it, otherwise
it will create a new one and add it to the cache.

To test our service, we'll want to assert that locations that have been
previously loaded are fetched from cache, while new locations are
created using the utility.